Record conversion to convert module
